[update] スクロール制御処理の定義を別ファイルに分離
