+++ src/background.ts
@@ -1,3 +1,5 @@
+import { startScroll } from './scrollControlCodes'
+
 // ロード終了時の処理
 chrome.tabs.onUpdated.addListener(function (tabId, info) {
   if (info.status === 'complete') {
@@ -15,6 +17,9 @@ chrome.tabs.onUpdated.addListener(function (tabId, info) {
         if (tabs.find((tab) => tab.id == object.currentTabId)) {
           // スクロール処理
           console.log('スクロール開始')
+          chrome.tabs.executeScript(tabId, {
+            code: `(${startScroll.toString()})()`,
+          })
         }
       })
     })
